+++ src/components/CardPreview/CardPreview.styles.ts
